--- src/main/java/zd/zero/waifu/motivator/plugin/alert/VisualMotivationFactory.kt
@@ -1,18 +1,0 @@
-package zd.zero.waifu.motivator.plugin.alert
-
-import com.intellij.openapi.project.Project
-import zd.zero.waifu.motivator.plugin.alert.notification.AlertConfiguration
-import zd.zero.waifu.motivator.plugin.player.WaifuSoundPlayerFactory
-
-object VisualMotivationFactory : WaifuMotivationFactory {
-    override fun constructMotivation(
-        project: Project,
-        motivationAsset: MotivationAsset,
-        config: AlertConfiguration
-    ): WaifuMotivation =
-        VisualMotivation(
-            VisualWaifuNotification(motivationAsset, project),
-            WaifuSoundPlayerFactory.createPlayer(motivationAsset.soundFileName),
-            config
-        )
-}
