+++ src/main/java/zd/zero/waifu/motivator/plugin/assets/MotivationAsset.kt
@@ -1,6 +1,7 @@
-package zd.zero.waifu.motivator.plugin.alert
+package zd.zero.waifu.motivator.plugin.assets
 
 import com.google.gson.annotations.SerializedName
+import zd.zero.waifu.motivator.plugin.alert.WaifuMotivatorAlertAssetCategory
 
 data class MotivationAsset(
     val title: String,
