+++ src/main/java/zd/zero/waifu/motivator/plugin/listeners/WaifuUnitTesterListenerImpl.java
@@ -2,13 +2,13 @@ package zd.zero.waifu.motivator.plugin.listeners;
 
 import com.intellij.openapi.project.Project;
 import zd.zero.waifu.motivator.plugin.alert.AlertAssetProvider;
-import zd.zero.waifu.motivator.plugin.alert.TextualMotivationFactory;
-import zd.zero.waifu.motivator.plugin.alert.VisualMotivationAssetProvider;
-import zd.zero.waifu.motivator.plugin.alert.VisualMotivationFactory;
-import zd.zero.waifu.motivator.plugin.alert.WaifuAssetCategory;
-import zd.zero.waifu.motivator.plugin.alert.WaifuMotivation;
+import zd.zero.waifu.motivator.plugin.motivation.TextualMotivationFactory;
+import zd.zero.waifu.motivator.plugin.assets.VisualMotivationAssetProvider;
+import zd.zero.waifu.motivator.plugin.motivation.VisualMotivationFactory;
+import zd.zero.waifu.motivator.plugin.assets.WaifuAssetCategory;
+import zd.zero.waifu.motivator.plugin.motivation.WaifuMotivation;
 import zd.zero.waifu.motivator.plugin.alert.WaifuMotivatorAlertAssetCategory;
-import zd.zero.waifu.motivator.plugin.alert.notification.AlertConfiguration;
+import zd.zero.waifu.motivator.plugin.alert.AlertConfiguration;
 import zd.zero.waifu.motivator.plugin.settings.WaifuMotivatorPluginState;
 import zd.zero.waifu.motivator.plugin.settings.WaifuMotivatorState;
