+++ src/main/java/zd/zero/waifu/motivator/plugin/motivation/VisualMotivation.kt
@@ -1,6 +1,7 @@
-package zd.zero.waifu.motivator.plugin.alert
+package zd.zero.waifu.motivator.plugin.motivation
 
-import zd.zero.waifu.motivator.plugin.alert.notification.AlertConfiguration
+import zd.zero.waifu.motivator.plugin.alert.WaifuNotification
+import zd.zero.waifu.motivator.plugin.alert.AlertConfiguration
 import zd.zero.waifu.motivator.plugin.player.WaifuSoundPlayer
 
 class VisualMotivation(
