+++ src/main/java/zd/zero/waifu/motivator/plugin/motivation/WaifuMotivationFactory.kt
@@ -1,7 +1,8 @@
-package zd.zero.waifu.motivator.plugin.alert
+package zd.zero.waifu.motivator.plugin.motivation
 
 import com.intellij.openapi.project.Project
-import zd.zero.waifu.motivator.plugin.alert.notification.AlertConfiguration
+import zd.zero.waifu.motivator.plugin.alert.AlertConfiguration
+import zd.zero.waifu.motivator.plugin.assets.MotivationAsset
 
 interface WaifuMotivationFactory {
     fun constructMotivation(project: Project, motivation: MotivationAsset, config: AlertConfiguration): WaifuMotivation
